Clean up routes.js with ES7 function bind syntax
Since passing a method as a callback messes with the this property, I
had used anonymous arrow methods in the routes to call the actual
controller methods. Another way would be to bind the methods to the
objects: instead of
  .get((req, res, next) => controller.method(req, res, next))
one could write
  .get(controller.method.bind(controller))
Which preserves the this property, and achieves the same without an
explicit intermediary function. The function bind syntax is a shorthand
for the same, so the bind can be written as:
  .get(::controller.method)

Since this is experimental syntax, it is being used through Babel.

https://github.com/zenparsing/es-function-bind